added handling not existing file input

+++ app/main.py
@@ -1,4 +1,6 @@
-# write your code here
+import os
+
+
 def copy_file(command: str) -> None:
     command_elems = command.split(" ")
     if len(command_elems) == 3:
@@ -6,7 +8,11 @@ def copy_file(command: str) -> None:
         filename = command_elems[1]
         copy_name = command_elems[2]
 
-    if len(command_elems) == 3 and command == "cp" and filename != copy_name:
+    if (
+        len(command_elems) == 3 and command == "cp"
+        and filename != copy_name
+        and os.path.exists(f"app/{filename}")
+    ):
         with open(filename, "r") as file_in, open(copy_name, "w") as file_out:
             original_text = file_in.read()
             file_out.write(original_text)
